Synchronization debugged

+++ MSS.WinMobile/MSS.WinMobile.Infrastructure.Repositoties/RouteTemplateRepository.cs
@@ -19,7 +19,7 @@ namespace MSS.WinMobile.Infrastructure.SqliteRepositoties
         private const string SaveQueryTemplate = "INSERT OR REPLACE INTO RouteTemplates (Id, DayOfWeek) VALUES ({0}, {1})";
         protected override string GetSaveQueryFor(RouteTemplate model)
         {
-            return string.Format(SaveQueryTemplate, model.Id, model.DayOfWeek);
+            return string.Format(SaveQueryTemplate, model.Id, (int)model.DayOfWeek);
         }
 
         private const string DeleteQueryTemplate = "DELETE FROM RouteTemplates WHERE Id = {0}";
